Fixes both the > 32768 port problem and the display of hosts on the summary screen.

+++ hoststbl.sql.php
@@ -5,8 +5,8 @@ CREATE TABLE IF NOT EXISTS `hosts` (
   `id` ".$primary_key.",
   `name` varchar(255) NOT NULL,
   `address` varchar(255) NOT NULL,
-  `port` smallint(6) NOT NULL DEFAULT '4028',
-  `mhash_desired` decimal(6,2) NOT NULL
+  `port` mediumint(6) NOT NULL DEFAULT '4028',
+  `mhash_desired` decimal(10,2) NOT NULL
 )".$table_props.";
 ";
